requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.1.0
requests==2.31.0
python-dateutil==2.8.2
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2
